monted context provider and isolated Stacked screens into component
